# api-nestjs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user/details": {
      "get": {
        "operationId": "userDetails",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Detalhes do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "role": {
                      "type": "string",
                      "example": "USER"
                    },
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "email": {
                      "type": "string",
                      "example": "usuario@email.com"
                    },
                    "firstName": {
                      "type": "string",
                      "example": "Nome"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "Sobrenome"
                    },
                    "whatsappNumber": {
                      "type": "string",
                      "example": "11999999999"
                    },
                    "verified": {
                      "type": "boolean",
                      "example": true
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2023-01-01T00:00:00.000Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2023-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter detalhes do usuário atual",
        "tags": [
          "Usuários"
        ]
      }
    },
    "/user/update": {
      "patch": {
        "operationId": "updateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuário atualizado."
                    },
                    "updatedFields": {
                      "type": "object",
                      "example": {
                        "fisrtName": "Novo",
                        "lastName": "Nome",
                        "whatsappNumber": "11999999999"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar usuário atual",
        "tags": [
          "Usuários"
        ]
      }
    },
    "/user/delete": {
      "delete": {
        "operationId": "deleteUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Usuário deletado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuário deletado."
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "role": {
                          "type": "string",
                          "example": "USER"
                        },
                        "id": {
                          "type": "number",
                          "example": 1
                        },
                        "email": {
                          "type": "string",
                          "example": "usuario@email.com"
                        },
                        "fisrtName": {
                          "type": "string",
                          "example": "Nome"
                        },
                        "lastName": {
                          "type": "string",
                          "example": "Sobrenome"
                        },
                        "whatsappNumber": {
                          "type": "string",
                          "example": "11999999999"
                        },
                        "imageUrl": {
                          "type": "string",
                          "example": "imageUrl"
                        },
                        "verified": {
                          "type": "boolean",
                          "example": true
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2023-01-01T00:00:00.000Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "example": "2023-01-01T00:00:00.000"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deletar usuário atual",
        "tags": [
          "Usuários"
        ]
      }
    },
    "/user/users": {
      "get": {
        "operationId": "getUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "example": [
                    {
                      "role": "USER",
                      "id": 1,
                      "email": "joao@email.com",
                      "firstName": "João",
                      "lastName": "Silva",
                      "whatsappNumber": "11999999999",
                      "imageUrl": "https://exemplo.com/joao.jpg",
                      "verified": true,
                      "createdAt": "2023-01-01T00:00:00.000Z",
                      "updatedAt": "2023-01-01T00:00:00.000Z"
                    },
                    {
                      "role": "ADMIN",
                      "id": 2,
                      "email": "maria@email.com",
                      "firstName": "Maria",
                      "lastName": "Santos",
                      "whatsappNumber": "11988888888",
                      "imageUrl": "https://exemplo.com/maria.jpg",
                      "verified": true,
                      "createdAt": "2023-01-02T00:00:00.000Z",
                      "updatedAt": "2023-01-02T00:00:00.000Z"
                    },
                    {
                      "role": "USER",
                      "id": 3,
                      "email": "pedro@email.com",
                      "firstName": "Pedro",
                      "lastName": "Oliveira",
                      "whatsappNumber": "11977777777",
                      "imageUrl": "https://exemplo.com/pedro.jpg",
                      "verified": false,
                      "createdAt": "2023-01-03T00:00:00.000Z",
                      "updatedAt": "2023-01-03T00:00:00.000Z"
                    }
                  ],
                  "items": {
                    "type": "object",
                    "properties": {
                      "role": {
                        "type": "string"
                      },
                      "id": {
                        "type": "number"
                      },
                      "email": {
                        "type": "string"
                      },
                      "firstName": {
                        "type": "string"
                      },
                      "lastName": {
                        "type": "string"
                      },
                      "whatsappNumber": {
                        "type": "string"
                      },
                      "imageUrl": {
                        "type": "string"
                      },
                      "verified": {
                        "type": "boolean"
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "updatedAt": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todos usuários (Admin)",
        "tags": [
          "Usuários"
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "operationId": "findUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "role": {
                      "type": "string",
                      "example": "USER"
                    },
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "email": {
                      "type": "string",
                      "example": "usuario@email.com"
                    },
                    "firstName": {
                      "type": "string",
                      "example": "Nome"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "Sobrenome"
                    },
                    "whatsappNumber": {
                      "type": "string",
                      "example": "11999999999",
                      "nullable": true
                    },
                    "imageUrl": {
                      "type": "string",
                      "example": "imageUrl"
                    },
                    "verified": {
                      "type": "boolean",
                      "example": true
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2024-11-30T16:14:20.100Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2024-11-30T17:02:17.452Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buscar usuário por ID (Admin)",
        "tags": [
          "Usuários"
        ]
      },
      "patch": {
        "operationId": "updateUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuário atualizado."
                    },
                    "updatedFields": {
                      "type": "object",
                      "example": {
                        "firstName": "novo",
                        "lastName": "nome",
                        "whatsappNumber": "11999999999"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar usuário por ID (Admin)",
        "tags": [
          "Usuários"
        ]
      },
      "delete": {
        "operationId": "deleteUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário deletado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuário deletado."
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "role": {
                          "type": "string",
                          "example": "USER"
                        },
                        "id": {
                          "type": "number",
                          "example": 1
                        },
                        "email": {
                          "type": "string",
                          "example": "usuario@email.com"
                        },
                        "fisrtName": {
                          "type": "string",
                          "example": "Nome"
                        },
                        "lastName": {
                          "type": "string",
                          "example": "Sobrenome"
                        },
                        "whatsappNumber": {
                          "type": "string",
                          "example": "11999999999"
                        },
                        "imageUrl": {
                          "type": "string",
                          "example": "imageUrl"
                        },
                        "verified": {
                          "type": "boolean",
                          "example": true
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2023-01-01T00:00:00.000Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "example": "2023-01-01T00:00:00.000"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deletar usuário por ID (Admin)",
        "tags": [
          "Usuários"
        ]
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email do usuário",
            "example": "usuario@email.com"
          },
          "firstName": {
            "type": "string",
            "description": "Primeiro nome do usuário",
            "example": "Nome"
          },
          "lastName": {
            "type": "string",
            "description": "Sobrenome do usuário",
            "example": "Sobrenome"
          },
          "whatsappNumber": {
            "type": "string",
            "description": "Número do WhatsApp",
            "example": "11999999999"
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName"
        ]
      },
...
    }
  }
}